pyproject.toml
[project]
name = "frc-design-app"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask[async]>=3.1.1",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-logging>=3.12.1",
    "google-cloud-storage>=3.4.1",
    "gunicorn>=23.0.0",
    "json5>=0.12.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "requests-oauthlib>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["onshape_api", "backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
